refactorization, initaial trajectory planning
